requirements.txt
streamlit
google-genai
pandas
openpyxl